Fix drag and drop in composer
Simplify how we handle the drag and drop. Now the drop is always handled
by the EContentEditor first and then if the data are processed by the
editor then it's handed to the EMsgComposer. Also the DOM operations
are now handled on the web extension side to save some DBus calls.